Added GraphQl Queries
Added Graphql Queries in gatsby.node.js and in the component genInfoTable.js for testing purposes.  gastby-node.js is the file that creates the pages for each user, and getInfoTable.js tests whether the system can access the username or not.  You will have to add an "ID" and "Path" field to currently access the data locally